requirements.txt
langchain_google_genai
langchain_community
python-dotenv
langchain-core
langchain-ollama
pydantic